Added a script for updating local APT caches from a master via SSH
First an rsync daemon needs to be started on localhost of the master
box, then an SSH tunnel needs to be created from the local box to the
listening port of the RSYNC daemon on the master box, and finally this
script needs to be run.

It should replace "apt-get update" and partly also
"apt-get upgrade --download".